+++ source/frontend/buttons/bbremovecompleters.js
@@ -32,7 +32,7 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				logicLayer.bounties.deleteSelectedBountyCompletions(bountyId, removedIds);
 				const poster = await logicLayer.hunters.findOneHunter(collectedInteraction.user.id, collectedInteraction.guild.id);
 				const company = await logicLayer.companies.findCompanyByPK(collectedInteraction.guildId);
-				buildBountyEmbed(bounty, collectedInteraction.guild, poster.getLevel(company.xpCoefficient), false, company, await logicLayer.bounties.findBountyCompletions(bountyId))
+				buildBountyEmbed(bounty, collectedInteraction.guild, poster.getLevel(company.xpCoefficient), false, company, await logicLayer.bounties.getHunterIdSet(bountyId))
 					.then(async embed => {
 						if (collectedInteraction.channel.archived) {
 							await collectedInteraction.channel.setArchived(false, "completers removed from bounty");