+++ source/frontend/commands/bounty/list.js
@@ -15,7 +15,7 @@ module.exports = new SubcommandWrapper("list", "List all of a hunter's open boun
 				const allHunters = await logicLayer.hunters.findCompanyHunters(interaction.guild.id);
 				const company = await logicLayer.companies.findCompanyByPK(interaction.guildId);
 				const companyLevel = company.getLevel(allHunters);
-				interaction.reply({ content: heading(`Evergreen Bounties on ${interaction.guild.name}`, 2), embeds: await Promise.all(existingBounties.map(async bounty => buildBountyEmbed(bounty, interaction.guild, companyLevel, false, company, await logicLayer.bounties.findBountyCompletions(bounty.id)))), flags: MessageFlags.Ephemeral });
+				interaction.reply({ content: heading(`Evergreen Bounties on ${interaction.guild.name}`, 2), embeds: await Promise.all(existingBounties.map(async bounty => buildBountyEmbed(bounty, interaction.guild, companyLevel, false, company, await logicLayer.bounties.getHunterIdSet(bounty.id)))), flags: MessageFlags.Ephemeral });
 			});
 		} else {
 			logicLayer.bounties.findOpenBounties(listUserId, interaction.guild.id).then(async existingBounties => {
@@ -25,7 +25,7 @@ module.exports = new SubcommandWrapper("list", "List all of a hunter's open boun
 				}
 				const hunter = await logicLayer.hunters.findOneHunter(listUserId, interaction.guild.id);
 				const company = await logicLayer.companies.findCompanyByPK(interaction.guildId);
-				interaction.reply({ content: heading(`${userMention(listUserId)}'s Bounties`, 2), embeds: await Promise.all(existingBounties.map(async bounty => buildBountyEmbed(bounty, interaction.guild, hunter.getLevel(company.xpCoefficient), false, company, await logicLayer.bounties.findBountyCompletions(bounty.id)))), flags: MessageFlags.Ephemeral });
+				interaction.reply({ content: heading(`${userMention(listUserId)}'s Bounties`, 2), embeds: await Promise.all(existingBounties.map(async bounty => buildBountyEmbed(bounty, interaction.guild, hunter.getLevel(company.xpCoefficient), false, company, await logicLayer.bounties.getHunterIdSet(bounty.id)))), flags: MessageFlags.Ephemeral });
 			});
 		}
 	}