+++ source/frontend/commands/bounty/record-turn-ins.js
@@ -4,7 +4,7 @@ const { listifyEN, commandMention, updatePosting, congratulationBuilder, disable
 const { timeConversion } = require("../../../shared");
 const { SKIP_INTERACTION_HANDLING, SAFE_DELIMITER } = require("../../../constants");
 
-module.exports = new SubcommandWrapper("verify-turn-in", "Verify up to 5 bounty hunters have turned in one of your bounties",
+module.exports = new SubcommandWrapper("record-turn-ins", "Record turn-ins of one of your bounties for up to 5 bounty hunters",
 	async function executeSubcommand(interaction, runMode, ...[logicLayer, poster]) {
 		const openBounties = await logicLayer.bounties.findOpenBounties(interaction.user.id, interaction.guild.id);
 		if (openBounties.length < 1) {
@@ -48,31 +48,28 @@ module.exports = new SubcommandWrapper("verify-turn-in", "Verify up to 5 bounty 
 						});
 						break;
 					case "hunters":
-						try {
-							let { bounty: returnedBounty, allCompleters, company, validatedCompleterIds, bannedIds } = await logicLayer.bounties.addCompleters(bounty, interaction.guild, [...collectedInteraction.members.values()], runMode);
-							const post = await updatePosting(collectedInteraction.guild, company, returnedBounty, poster.getLevel(company.xpCoefficient), allCompleters);
-							const sentences = [];
-							if (validatedCompleterIds.length > 0) {
-								sentences.push(`The following bounty hunters' turn-ins of ${bold(returnedBounty.title)} have been recorded: ${listifyEN(validatedCompleterIds.map(id => userMention(id)))}`);
-								if (post) {
-									post.channel.send({ content: `${listifyEN(validatedCompleterIds.map(id => userMention(id)))} ${validatedCompleterIds.length === 1 ? "has" : "have"} turned in this bounty! ${congratulationBuilder()}!` });
-								}
+						const { eligibleTurnInIds, newTurnInIds, bannedTurnInIds } = await logicLayer.bounties.checkTurnInEligibility(bounty, [...collectedInteraction.members.values()], runMode);
+						const sentences = [];
+						if (newTurnInIds.size < 1) {
+							sentences.push("No new turn-ins were able to be recorded. You cannot credit yourself or bots for your own bounties.");
+						} else {
+							await logicLayer.bounties.bulkCreateCompletions(bounty.id, bounty.companyId, Array.from(eligibleTurnInIds), null);
+							const company = await logicLayer.companies.findCompanyByPK(bounty.companyId);
+							const newTurnInList = listifyEN(Array.from(newTurnInIds.values().map(id => userMention(id))));
+							sentences.push(`Turn-ins of ${bold(bounty.title)} have been recorded for the following hunters: ${newTurnInList}`);
+							const post = await updatePosting(collectedInteraction.guild, company, bounty, poster.getLevel(company.xpCoefficient), eligibleTurnInIds);
+							if (post) {
+								post.channel.send({ content: `${newTurnInList} ${newTurnInIds.size === 1 ? "has" : "have"} turned in this bounty! ${congratulationBuilder()}!` });
 							}
-							if (bannedIds.length > 0) {
-								sentences.push(`The following users were skipped due to currently being banned from using BountyBot: ${listifyEN(bannedIds.map(id => userMention(id)))}`);
-							}
-							collectedInteraction.update({
-								content: sentences.join("\n\n"),
-								components: []
-							});
-						} catch (e) {
-							if (typeof e !== 'string') {
-								console.error(e);
-							} else {
-								collectedInteraction.update({ content: e, components: [] });
-							}
-							return;
 						}
+
+						if (bannedTurnInIds.size > 0) {
+							sentences.push(`The following users were skipped due to currently being banned from using BountyBot: ${listifyEN(Array.from(bannedTurnInIds.values().map(id => userMention(id))))}`);
+						}
+						collectedInteraction.update({
+							content: sentences.join("\n\n"),
+							components: []
+						});
 						break;
 				}
 			})