+++ source/frontend/commands/evergreen/edit.js
@@ -104,7 +104,7 @@ module.exports = new SubcommandWrapper("edit", "Change the name, description, or
 				const allHunters = await logicLayer.hunters.findCompanyHunters(modalSubmission.guild.id);
 				const currentCompanyLevel = company.getLevel(allHunters);
 				if (company.bountyBoardId) {
-					const embeds = await Promise.all(openBounties.map(bounty => buildBountyEmbed(bounty, modalSubmission.guild, currentCompanyLevel, false, company, [])));
+					const embeds = await Promise.all(openBounties.map(bounty => buildBountyEmbed(bounty, modalSubmission.guild, currentCompanyLevel, false, company, new Set())));
 					const bountyBoard = await modalSubmission.guild.channels.fetch(company.bountyBoardId);
 					bountyBoard.threads.fetch(company.evergreenThreadId).then(async thread => {
 						const message = await thread.fetchStarterMessage();
@@ -112,7 +112,7 @@ module.exports = new SubcommandWrapper("edit", "Change the name, description, or
 					});
 				}
 
-				const bountyEmbed = await buildBountyEmbed(selectedBounty, modalSubmission.guild, currentCompanyLevel, false, company, []);
+				const bountyEmbed = await buildBountyEmbed(selectedBounty, modalSubmission.guild, currentCompanyLevel, false, company, new Set());
 				modalSubmission.reply({ content: "Here's the embed for the newly edited evergreen bounty:", embeds: [bountyEmbed], flags: MessageFlags.Ephemeral });
 			});
 		}).catch(error => {